HotelRestaurantRetailsProject/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
awsebcli==3.20.6
botocore==1.29.99
cement==2.8.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
cryptography==40.0.2
decouple==0.0.7
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
dj-database-url==2.0.0
dj-rest-auth==4.0.1
Django==4.2
django-heroku==0.3.1
django-import-export==3.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-nested-routers==0.93.4
et-xmlfile==1.1.0
filelock==3.8.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
imageio==2.28.1
import-export==0.3.1
jmespath==1.0.1
keras==2.12.0
keras-facenet==0.3.2
MarkupPy==1.14
mtcnn==0.1.1
multidict==6.0.4
numpy==1.24.3
odfpy==1.4.1
opencv-python==4.7.0.72
openpyxl==3.1.2
pathspec==0.10.1
Pillow==9.5.0
platformdirs==2.5.4
psycopg2==2.9.6
pycparser==2.21
PyJWT==1.7.1
pypiwin32==223
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
pytz==2023.3
pywin32==306
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
sendgrid==6.10.0
six==1.14.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tablib==3.5.0
termcolor==1.1.0
twilio==8.3.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.17.0
virtualenvwrapper-win==1.2.7
wcwidth==0.1.9
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
